requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
requests
streamlit
streamlit-calendar
pyyaml
